Create blackjack.c
This is the start for the blackjack simulator in the language c, I am not as experienced with c as I am with python, thus this should take a little bit longer. I might try to also remake it in java and then time all of the files to see which one runs the quickest.